+++ packages/graphql_server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 6.5.0
+
+* Require Dart >= 3.9
+
+## 6.4.1
+
+* Fix analyzer warnings
+
 ## 6.4.0
 
 * Require Dart >= 3.8